python-service/pyproject.toml
[project]
name = "python-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "prometheus-fastapi-instrumentator>=7.1.0",
    "python-dotenv>=1.1.0",
    "uvicorn[standard]>=0.34.3",
]
